// asr.webapp.v2/src/app/modules/financial/trial-balance/details/dialog.service.ts
import { Injectable } from '@angular/core';
import { MatDialog, MatDialogRef } from '@angular/material/dialog';
import { merge } from 'lodash-es';
import { TrialBalanceDetailDialogConfig } from './dialog.types';
import { TrialBalanceDetailDialogComponent } from './dialog/dialog.component';

@Injectable()
export class DialogService {
    private _defaultConfig: TrialBalanceDetailDialogConfig = {
        requestId: 0,
        requestTypeId: 0,
        title: 'Confirm action',
        message: 'Are you sure you want to confirm this action?',
        icon: {
            show: true,
            name: 'heroicons_outline:information-circle',
            color: 'info'
        },
        actions: {
            confirm: {
                show: true,
                label: 'Confirm',
                color: 'warn'
            },
            cancel: {
                show: true,
                label: 'Cancel'
            }
        },
        dismissible: true
    };

    /**
     * Constructor
     */
    constructor(
        private _matDialog: MatDialog
    ) {
    }

    // -----------------------------------------------------------------------------------------------------
    // @ Public methods
    // -----------------------------------------------------------------------------------------------------

    open(config: TrialBalanceDetailDialogConfig = {}): MatDialogRef<TrialBalanceDetailDialogComponent> {
        // Merge the user config with the default config
        const userConfig = merge({}, this._defaultConfig, config);

        // Open the dialog
        return this._matDialog.open(TrialBalanceDetailDialogComponent, {
            autoFocus: false,
            disableClose: !userConfig.dismissible,
            data: userConfig
        });
    }
}
